Remove code duplication on Unit test
Code duplication (aka having the same thing all over the source) is not
a good practice because it lowers maintainability and code refactoring
would be a stressful task.

If the same lines of code are used, it should be it's own function rather
than copy-pasting the same lines of code all over the project.

I've also fixed a couple of grammar problems on the comments, like using -s
suffix on a verb that's lead by a first person subject.